Time: 0 ms (100.00%) | Memory: 8.6 MB (64.95%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int original=x;
-        long reversed=0;
+        long long reversed=0;
         if(x<0) return false;
         while(x!=0){
             int rem=x%10;
